pyproject.toml
[tool.poetry]
name = "jneqsim"
# Set in CI upon publishing
version = "0.0.0"
readme = "README.md"
description = "Python bindings for the Java library NeqSim (via JPype), with python stubs"
authors = ["Jørgen <jenge@equinor.com>","Stig Oskar <stoo@equinor.com>"]
maintainers = [
  "Jørgen <jenge@equinor.com>",
  "Stig Oskar <stoo@equinor.com>",
]
homepage = "https://github.com/equinor/jneqsim"
repository = "https://github.com/equinor/jneqsim"
keywords = ["neqsim"]
license = "Apache-2.0"
include = ["jneqsim/neqsim.jar", "docs/"]

[tool.poetry.dependencies]
python = "^3.12"
JPype1 = "^1.4.1"

[tool.poetry.group.dev.dependencies]
stubgenj = "^0.2.12"
pytest = "^7.4.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
)/
'''

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "S",
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # automatically upgrade syntax for newer versions of the language
    "RUF", # ruff specific rules
]

target-version = "py312"

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"] # Ignore `E402` (import violations) in all `__init__.py` files
"**/tests/*" = ["E402", "S101"]
